requirements.txt
python-dotenv
azure-ai-formrecognizer
opencv-python
imutils
requests
matplotlib
numpy
